% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_ts_code.R
\name{edit_ts_code}
\alias{edit_ts_code}
\title{Edit timeseries codes}
\usage{
edit_ts_code(ts_code_file, use_browser = TRUE)
}
\arguments{
\item{ts_code_file}{a filename with timeseries codes. This file does not have
to exist yet. If the file does exists, then it should be an rds file 
containing a \code{table_code_collection} object.}

\item{use_browser}{if \code{TRUE}, then display the graphical user interface
in the browser. Otherwise the RStudio viewer is used.}
}
\description{
Edit timeseries codes
}
