% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_tables_from_table.R
\name{fill_tables_from_table}
\alias{fill_tables_from_table}
\title{Fill tables with values from another table.}
\usage{
fill_tables_from_table(tscodes, ids, basis_id)
}
\arguments{
\item{tscodes}{a \code{ts_code} object. This object can be created
and modified with function \code{\link{edit_ts_code}}, which starts a Shiny
app}

\item{ids}{ids of tables to be filled from the second tabel}

\item{basis_id}{id of the table whose values will be used to fill
the first tables.}
}
\description{
Fill tables with values from another table.
}
