% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_modified.R
\name{get_last_modified}
\alias{get_last_modified}
\title{Return the last time that the timeseries code for one or more tables has been
modified.}
\usage{
get_last_modified(id, ts_code)
}
\arguments{
\item{id}{table id a character vector with CBS table identifiers}

\item{ts_code}{a \code{ts_code} object. This object can be created 
with a Shiny app that you start with function
\code{\link{edit_ts_code}}}
}
\value{
the last time that the code for the CBS tables have been modified.
}
\description{
Return the last time that the timeseries code for one or more tables has been
modified.
}
