% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts.R
\name{get_ts}
\alias{get_ts}
\title{Return time indices}
\usage{
get_ts(id, table_code_collection, download, raw_cbs_dir = "raw_cbs_data")
}
\arguments{
\item{id}{table id}

\item{table_code_collection}{a \code{table_code_collection} object}

\item{download}{If \code{TRUE}, data are downloaded, otherwise the data
is read from the previously downloaded data in directory \code{raw_cbs_dir}}

\item{raw_cbs_dir}{directory where the raw downloaded data are stored}
}
\description{
Return time indices
}
