% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_code.R
\name{join_tables}
\alias{join_tables}
\title{Join two  \code{ts_code} objects.}
\usage{
join_tables(tsc1, tsc2)
}
\arguments{
\item{tsc1}{The first  \code{ts_code} object.}

\item{tsc2}{The second  \code{ts_code} object.}
}
\value{
A new \code{ts_code} object with the union of the tables in \code{ts_code}
objects \code{tsc1} and \code{tscs2}.
}
\description{
The two objects should not have common table ids. Use function
\code{\link{select_tables}} to select the tables that should be joined.
}
\examples{
ts_code_1 <- select_tables(ts_code_example, ids =  c("70076ned", "7116shfo"))
ts_code_2 <- select_tables(ts_code_example, ids =  "00372")
join_tables(ts_code_1, ts_code_2)
}
\seealso{
\code{select_tables}
}
