% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ts_code_xlsx.R
\name{read_ts_code_xlsx}
\alias{read_ts_code_xlsx}
\title{Read timeseries coding from an Excel file.}
\usage{
read_ts_code_xlsx(xlsx_files)
}
\arguments{
\item{xlsx_files}{a named character vector with the names of one or more Excel
files with timeseries information. The names are the table ids as
defined by the cbs.}
}
\description{
The format of the Excel file is the old-style timeseries coding.
}
