% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_code.R
\name{select_tables}
\alias{select_tables}
\title{Select tables from a \code{ts_code} object}
\usage{
select_tables(tsc, ids)
}
\arguments{
\item{tsc}{A \code{ts_code} object.}

\item{ids}{A character vector with table identifiers or a numeric vector
with the indices of the tables.}
}
\value{
A new \code{ts_code} object with a selection of tables.
}
\description{
Select tables from a \code{ts_code} object
}
\examples{
select_tables(ts_code_example, ids =  c("70076ned", "7116shfo"))

# this expression gives the same result:
select_tables(ts_code_example, ids = c(2,3))
}
\seealso{
\code{join_tables}
}
