% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_code_example.R
\docType{data}
\name{ts_code_example}
\alias{ts_code_example}
\title{An example \code{ts_code} object}
\format{
An object of class \code{ts_code} of length 3.
}
\usage{
ts_code_example
}
\description{
A \code{ts_code} object containing timeseries coding for three CBS tables.
This object is used for demonstrating several functions of
package \code{cbsots}.
}
\keyword{datasets}
