% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tables.R
\name{update_tables}
\alias{update_tables}
\title{Update tables}
\usage{
update_tables(tscodes, ids, base_url = NULL)
}
\arguments{
\item{tscodes}{a \code{ts_code} object. This object can be created
and modified with function \code{\link{edit_ts_code}}, which starts a Shiny
app}

\item{ids}{ids of tables to be updated}

\item{base_url}{optionally specify a different server. Useful for third party
data services implementing the same protocol.}
}
\description{
This function updates one or more tables in a table_code object.
}
