% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_ts_xlsx.R
\name{write_table_ts_xlsx}
\alias{write_table_ts_xlsx}
\title{Writes the timeseries returned by function \code{\link{get_ts}} to
an Excel file.}
\usage{
write_table_ts_xlsx(x, file, rowwise = TRUE, ...)
}
\arguments{
\item{x}{the \code{table_ts} object returned by function \code{\link{get_ts}}}

\item{file}{a filename}

\item{rowwise}{a logical value: should the timeseries be written rowwise?}

\item{...}{other arguments passed to 
\code{\link[regts:write_ts_xlsx-slash-write_ts_sheet]{write_ts_xlsx}}}
}
\description{
Writes the timeseries returned by function \code{\link{get_ts}} to
an Excel file.
}
