% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_ts_code.R
\name{edit_ts_code}
\alias{edit_ts_code}
\title{Edit timeseries codes}
\usage{
edit_ts_code(ts_code_file, use_browser = TRUE, debug = FALSE)
}
\arguments{
\item{ts_code_file}{the name of a file where the timeseries coding is
stored.  This file does not have to exist yet. If it does exist, then it 
should be an \code{rds} file containing a \code{ts_code} object.}

\item{use_browser}{if \code{TRUE}, then display the graphical user interface
in the browser. Otherwise the RStudio viewer is used.}

\item{debug}{a logical. If \code{TRUE}, then use the debugging mode
(only for developpers)}
}
\description{
Edit timeseries codes
}
