% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_ts_code.R
\name{edit_ts_code}
\alias{edit_ts_code}
\title{Edit timeseries codes}
\usage{
edit_ts_code(ts_code_file, use_browser = TRUE, browser, debug = FALSE,
  base_url = NULL)
}
\arguments{
\item{ts_code_file}{the name of a file where the timeseries coding is
stored.  This file does not have to exist yet. If it does exist, then it
should be an \code{rds} file containing a \code{ts_code} object.}

\item{use_browser}{if \code{TRUE} (the default), then display the graphical
user interface in the browser. Otherwise the RStudio viewer is used.}

\item{browser}{a character vector specifying the path of the browser. Specify
\code{"default"} to use the default browser. The approach used when this
argument has not been specified depends on the operating system. For non-Windows
operating systems, the default browser is also used if argument
\code{browser} has not been specified. For Windows, a different approach is
used because the Shiny app does not work well in the Internet Explorer.
The function tries to find the location of Chrome or FireFox and if the search
is succesful then this browser is used. Otherwise an error is issued.}

\item{debug}{a logical. If \code{TRUE}, then use the debugging mode
(only for developers)}

\item{base_url}{optionally specify a different server. Useful for third party
data services implementing the same protocol.}
}
\description{
Edit timeseries codes
}
