% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_tables_from_table.R
\name{fill_tables_from_table}
\alias{fill_tables_from_table}
\title{Fill tables with values from another table.}
\usage{
fill_tables_from_table(ts_code, ids, base_id, base_url = NULL)
}
\arguments{
\item{ts_code}{a \code{ts_code} object. This object can be created
and modified with function \code{\link{edit_ts_code}}, which starts a Shiny
app}

\item{ids}{ids of tables to be filled from the base tabel}

\item{base_id}{the id of the base table}

\item{base_url}{optionally specify a different server. Useful for third party
data services implementing the same protocol.}
}
\description{
This function fills in the Select and Code fields from a base table based
on common Keys or Titles. The function tries to find matching keys or titles
and updates the corresponding Select en Code fields.
The
function creates a match report in directory \code{match_reoport}
for each table and dimension. The match report is an Excel file with a name
composed of the table id of the base table and the new
table, and the and the dimension, e.g.
\code{83186NED_83361NED_TypeZelfstandige.xlsx}.
This function is still experimental and you should always check the
results carefully.
}
