% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tables.R
\name{update_tables}
\alias{update_tables}
\title{Update tables}
\usage{
update_tables(ts_code, ids, base_url = NULL)
}
\arguments{
\item{ts_code}{a \code{ts_code} object. This object can be created
and modified with function \code{\link{edit_ts_code}}, which starts a Shiny
app}

\item{ids}{ids of tables to be updated}

\item{base_url}{optionally specify a different server. Useful for third party
data services implementing the same protocol.}
}
\description{
This function updates one or more tables in a \code{ts_code} object.
The function downloads the latest meta data from the CBS website
and tries to find  matching keys or titles in the old and new table. The
function creates a match report in directory \code{match_reoport}
for each table and dimension. The match report is an Excel file with a name
composed of the table id and dimension, e.g. \code{81810NED_Topic.xlsx}.
This function is still experimental and you should always check the results
carefully.
}
\examples{
\dontrun{
update_tables(ts_code_example)
}
}
